quantity update working
